Depend on elasticsearch-parent
To simplify plugins maintenance and provide more value in the future, we are starting to build an `elasticsearch-parent` project.
This commit is the first step for this plugin to depend on this new `pom` maven project.

(cherry picked from commit 0c910898756f52b67cdb8e6f110f292b7a0ed3eb)